Refactor GCSE controller into type, details and review.
