+++ app/controllers/candidate_interface/gcse/details_controller.rb
@@ -1,42 +1,18 @@
 module CandidateInterface
-  class GcseDetailsController < CandidateInterfaceController
+  class Gcse::DetailsController < CandidateInterfaceController
     before_action :set_subject
 
-    # 1th step - Edit qualification type
-    def edit_type
-      @application_qualification = GcseQualificationTypeForm.new(
-        subject: subject_param,
-        level: ApplicationQualification.levels[:gcse],
-      )
-    end
-
-    def update_type
-      @application_qualification = GcseQualificationTypeForm.new(qualification_type: qualification_type_param,
-                                                                 subject: subject_param,
-                                                                 level: ApplicationQualification.levels[:gcse])
-
-      application_form = current_candidate.current_application
-
-      if @application_qualification.save_base(application_form)
-        redirect_to candidate_interface_gcse_details_edit_details_path
-      else
-        render :edit_type
-      end
-    end
-
     # 2nd step - Edit grade and award year
-    def edit_details
+    def edit
       application_qualification = ApplicationQualification.last
 
       @application_qualification = GcseQualificationDetailsForm.new(
         grade: application_qualification.grade,
         award_year: application_qualification.award_year,
       )
-
-      render :edit_details
     end
 
-    def update_details
+    def update
       current_application_qualification = ApplicationQualification.last
 
       details_form = GcseQualificationDetailsForm.new(
@@ -69,10 +45,6 @@ module CandidateInterface
 
     def subject_param
       params.require(:subject)
-    end
-
-    def qualification_type_param
-      (params[:candidate_interface_gcse_qualification_type_form] || {}).fetch(:qualification_type, '')
     end
   end
 end
