+++ app/controllers/candidate_interface/gcse/review_controller.rb
@@ -1,0 +1,19 @@
+module CandidateInterface
+  class Gcse::ReviewController < CandidateInterfaceController
+    before_action :set_subject
+
+    def show
+      @application_qualification = ApplicationQualification.last
+    end
+
+  private
+
+    def set_subject
+      @subject = subject_param
+    end
+
+    def subject_param
+      params.require(:subject)
+    end
+  end
+end
