+++ app/views/candidate_interface/gcse/details/edit.html.erb
