+++ app/views/candidate_interface/gcse/type/edit.html.erb
